Add acceptance tests
Because key-chord-multiple relies on quite subtle behavior at times,
realistic acceptance tests are a nice thing to have.

The best I could come up with is using the expect program to "fake"
keyboard input that is sent to Emacs. I also tried pushing events the
unread-command-events list directly in elisp, but couldn't immediately
figure out a way how to force Emacs to handle all the events in that
list. Using expect is nicer anyway, as it is provides a more realistic
simulation.

The tests can be run locally by simply running

./tests/run-tests.sh

Internally the tests work by defining key-chords that trigger Emacs to
write some text to a file somehow. Two functions have been provided in
elisp: write-current-line-to-file and write-string-to-file. When calling
one of these functions a new file with an incremented file name will be
created with the output in it.

The expected test results are provided in tests/test_results, at the end
of the acceptance tests a recursive diff will be executed to figure out
if everything worked as expected.