Allow binding key-chords to lambdas
In the keymap the lambda is not inside of a list. So when traversing the
keymap we have to explicitly test whether or not what is left is a
function or not.

We also tested whether or not an element was a symbol before attempting
to execute it. This was wrong as a lambda is not a symbol but a
function.